Logique service de fetchCourseStructure transférée Logique transférée vers le fichier functions-said Les imports sont optimisés
